Reordered code to update fusion tables; added people counter node reboot protection

+++ fusiontables-cmdline-sample/src/main/java/it/uniurb/disbef/virtualsense/basestation/FusionTableGlobalPeopleRecord.java
@@ -7,5 +7,5 @@ package it.uniurb.disbef.virtualsense.basestation;
 public class FusionTableGlobalPeopleRecord {
   public short out;
   public short in;    
-  public short inside;
+  public int inside;
 }
